% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{assess_transfer_curve}
\alias{assess_transfer_curve}
\title{Assess quality of transfer curve}
\usage{
assess_transfer_curve(x, type = "p")
}
\arguments{
\item{x}{A data frame of I-V data for one sweep.}

\item{type}{A character (\code{"p"} or \code{"n"})}
}
\value{
A data frame with four rows and three columns: \code{term}, \code{value},
\code{flag}. Four QC terms are considered: \code{on_off_ratio}, \code{on_off_rng},
\code{ID_max}, and \code{ID_IG_ratio}.
}
\description{
For a transfer curve (sqrt(I_D) vs. V_G), calculate quality measures about
on-off ratio, on-off range, magnitude of drain current, and ratio of
average drain current to average gate current, and flag abnormal instances
based on pre-defined thresholds.
}
