% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{averaged_diff}
\alias{averaged_diff}
\title{Average of numerical difference}
\usage{
averaged_diff(x, y, n = 5)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector.}

\item{n}{An integer to indicate the number of points to be averaged.}
}
\value{
A numeric vector of the averaged derivative.
}
\description{
A wrapper of \code{num_diff()} and \code{moving_ave} to find smoothed derivative
}
