% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{characterize_transfer_curve}
\alias{characterize_transfer_curve}
\title{Characterize transfer curve}
\usage{
characterize_transfer_curve(x)
}
\arguments{
\item{x}{A data frame of I-V data for one sweep.}
}
\value{
A data frame with one row and four columns: \code{gm_max}, \code{V_thresh},
\code{slope}, \code{auc}.
}
\description{
Based on gate voltage, drain current, and transconductance, calculate
maximum transconductance, threshold voltage, slope of the transfer curve
(sqrt(I_D) vs. V_G), and area under the transfer curve.
}
