% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_extension}
\alias{extract_extension}
\title{Extract file extension from filename}
\usage{
extract_extension(path)
}
\arguments{
\item{path}{A string of path to a file.}
}
\value{
File extension extracted from the input string.
}
\description{
Extract file extension from filename
}
\examples{
extract_extension("my_file.ext")
}
