% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{index_sweep}
\alias{index_sweep}
\title{Check and annotate the ordering of the input}
\usage{
index_sweep(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A list of two indexing vectors named \code{fwd} and \code{bwd}.
}
\description{
The function can be used to check the sweeping mode with V_gate, which is
supposed to be in ascending order (forward sweeping) and then descending
order (backward sweep). It returns a list of two indexing vectors named
\code{fwd} and \code{bwd}.
}
\examples{
index_sweep(c(seq(0, 0.5, by = 0.05), seq(0.45, 0, by = -0.05)))
}
