% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-classes.R
\name{is_OTAnalysis}
\alias{is_OTAnalysis}
\title{Test if the object is an \code{OTAnalysis}}
\usage{
is_OTAnalysis(x)
}
\arguments{
\item{x}{An object}
}
\value{
TRUE if the object inherits from the \code{OTAnalysis} class.
}
\description{
This function returns TRUE for \code{OTAnalysis}, and FALSE for all other objects.
}
