% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{moving_ave}
\alias{moving_ave}
\title{Moving average}
\usage{
moving_ave(x, n = 5)
}
\arguments{
\item{x}{A numeric vector.}

\item{n}{An integer to indicate the number of points to be averaged.}
}
\value{
A numeric vector of the averaged result.
}
\description{
Moving average
}
\examples{
moving_ave(1:20, n = 3)
}
