% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{num_diff}
\alias{num_diff}
\title{Numerical differentiation}
\usage{
num_diff(x, y)
}
\arguments{
\item{x}{A numeric vector \eqn{x}.}

\item{y}{A numeric vector \eqn{y=f(x)}.}
}
\value{
A numeric vector for the approximate derivative \eqn{dy/dx}.
}
\description{
Numerical differentiation to approximate the derivative \eqn{dy/dx}.
}
\examples{
set.seed(1)
num_diff(1:10, 3 * (1:10) + rnorm(10))
}
