% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{OT_analyze}
\alias{OT_analyze}
\title{Processing and analysis of \code{OTDataSet}}
\usage{
OT_analyze(x, n_ave = 5, ignore_outlier = TRUE)
}
\arguments{
\item{x}{An \code{OTDataSet} object.}

\item{n_ave}{An integer defining number of points for moving average.}

\item{ignore_outlier}{A logical value specifying whether outliers should be
ignored. Default is \code{TRUE}.}
}
\value{
An \code{OTAnalysis} object.
}
\description{
Based on the input I-V data restored in an \code{OTDataSet} object, separate into
sweep-level I-V data (one for forward sweep, one for backward sweep),
process and analyze the sweep-level data for quality assessment and
parameter estimation, and summarize each transistor based on its sweep-level
estimates.
}
