% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{OT_export_curves}
\alias{OT_export_curves}
\title{Export multiple organic transistor measurement curves to a PDF}
\usage{
OT_export_curves(
  data,
  x_var = "V_gate",
  y_var = "abs_I_drain",
  y2_var = NULL,
  sweep = "both",
  x_lab = x_var,
  y_lab = y_var,
  y2_lab = y2_var,
  y_trans = "identity"
)
}
\arguments{
\item{data}{An \code{OTAnalysis} object.}

\item{x_var}{A character.}

\item{y_var}{A character.}

\item{y2_var}{An optional character.}

\item{sweep}{A character with three possible values (\code{"fwd"}, \code{"bwd"}, or
\code{"both"} (default)).}

\item{x_lab}{A character or expression.}

\item{y_lab}{A character or expression.}

\item{y2_lab}{An optional character or expression.}

\item{y_trans}{A character.}
}
\value{
A named integer giving the number and name of the new active device
(after the specified device has been shut down).
}
\description{
\code{OT_export_curves()} is a wrapper of \code{OT_plot_curve()}, creating and saving
multiple organic transistor measurement curves to a PDF.
}
