% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{OT_export_results}
\alias{OT_export_results}
\title{Export and save analysis results to a CSV}
\usage{
OT_export_results(x, file = NULL, level = "group", by_var = ".id")
}
\arguments{
\item{x}{An \code{OTAnalysis} object.}

\item{file}{CSV file to write to.}

\item{level}{A character with two possible values (\code{"group"} (default) or
\code{"OT"}).}

\item{by_var}{Character vector for variables in metadata to define groups
for summarization.}
}
\value{
The result data frame is returned invisibly.
}
\description{
This function exports the group- or transistor-level results to a CSV. The
transistor-level results are available in a valid \code{OTAnalysis} input. The
function calls the \code{OT_results()} function when the group-level results are
exported.
}
