% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{OT_flag_display}
\alias{OT_flag_display}
\title{Display flagged transistors and sweeps}
\usage{
OT_flag_display(x)
}
\arguments{
\item{x}{An \code{OTAnalysis} object.}
}
\value{
A data frame of sweep-level estimates for flagged instances.
}
\description{
Display flagged transistors and sweeps
}
