% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{OT_flag_modify}
\alias{OT_flag_modify}
\title{Modify (flag or unflag) transistors and sweeps}
\usage{
OT_flag_modify(x, mod)
}
\arguments{
\item{x}{An \code{OTAnalysis} object.}

\item{mod}{A data frame of three columns: \code{.id}, \code{sweep}, \code{flag}.}
}
\value{
An \code{OTAnalysis} object.
}
\description{
Modify (flag or unflag) transistors and sweeps
}
