% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{OT_plot_curve}
\alias{OT_plot_curve}
\title{Plot organic transistor measurement curve(s)}
\usage{
OT_plot_curve(
  data,
  x_var = "V_gate",
  y_var = "abs_I_drain",
  y2_var = NULL,
  group = NULL,
  sweep = "both",
  .id = NULL,
  x_lab = x_var,
  y_lab = y_var,
  y2_lab = y2_var,
  y_trans = "identity"
)
}
\arguments{
\item{data}{An \code{OTAnalysis} object.}

\item{x_var}{A character.}

\item{y_var}{A character.}

\item{y2_var}{An optional character.}

\item{group}{An optional character.}

\item{sweep}{A character with three possible values (\code{"fwd"}, \code{"bwd"}, or
\code{"both"} (default)).}

\item{.id}{An optional integer.}

\item{x_lab}{A character or expression.}

\item{y_lab}{A character or expression.}

\item{y2_lab}{An optional character or expression.}

\item{y_trans}{A character.}
}
\value{
A \code{ggplot} object or a list of \code{ggplot} objects.
}
\description{
\code{OT_plot_curve()} returns a \code{ggplot} object or a list of \code{ggplot} objects
specifying measurement curve(s) for one or multiple organic transistors.
}
\details{
By default, the function plots the transfer characteristic curve for each
transistor, which is useful in most cases. Specifically, the \code{x_var} argument
is set to \code{"V_gate"} and the \code{y_var} argument is set to \code{"abs_I_drain"}.
The plot may include an optional second curve, defined based on another
variable for an additional y-axis on the right of the plot (the \code{y2_var}
argument). The same \code{x_var} will be used for the second curve.
Transconductance (\code{"gm"}) and gate current (\code{"I_gate"}) are common choices
for \code{y2_var}. Default is to create such curves for all the transistors, and
return a list of \code{ggplot} objects, one for each transistor. Alternatively,
user can choose to plot just one transistor, indexed by the \code{.id} argument.
In that case, a \code{ggplot} object will be returned by the function. Another
possibility is to have all the curves in one plot, colored based on a
specified grouping variable (argument \code{group}) that should be part of the
metadata. By default, both \code{y2_var} and \code{group} are set to \code{NULL}. It is not
allowed to specify both arguments at the same time.
}
