% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{OT_plot_param}
\alias{OT_plot_param}
\title{Plot estimates of organic transistor parameters}
\usage{
OT_plot_param(
  data,
  x_var = NULL,
  y_var = character(),
  group = NULL,
  level = "sweep",
  use_raw_x = FALSE,
  ignore_outlier = FALSE,
  x_lab = x_var,
  y_lab = y_var
)
}
\arguments{
\item{data}{An \code{OTAnalysis} object.}

\item{x_var}{An optional character.}

\item{y_var}{A character.}

\item{group}{An optional character.}

\item{level}{A character with two possible values (\code{"sweep"} (default) or
\code{"OT"}).}

\item{use_raw_x}{A logical value.}

\item{ignore_outlier}{A logical value.}

\item{x_lab}{A character or expression.}

\item{y_lab}{A character or expression.}
}
\value{
A \code{ggplot} object.
}
\description{
\code{OT_plot_param()} returns a \code{ggplot} object for parameter estimates of the
analyzed organic transistors.
}
