% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{OT_results}
\alias{OT_results}
\title{Summarize analysis results per defined group}
\usage{
OT_results(x, by_var = ".id")
}
\arguments{
\item{x}{An \code{OTAnalysis} object.}

\item{by_var}{Character vector for variables in metadata to define groups
for summarization.}
}
\value{
A data frame.
}
\description{
For each group defined based on variables in the metadata, summarize in
terms of number of transistors (\code{n}), number of flagged transistors
(\code{n_flagged}), mean and standard deviation of
maximum transconductance (\code{gm_max_mean} and \code{gm_max_sd}), and
shift of threshold voltage (\code{V_thresh_shift_mean} and \code{V_thresh_shift_sd}).
}
