% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-classes.R
\docType{class}
\name{new_OTAnalysis}
\alias{new_OTAnalysis}
\alias{OTAnalysis}
\alias{OTAnalysis-class}
\alias{validate_OTAnalysis}
\title{OTAnalysis constructor and validator}
\usage{
new_OTAnalysis(
  IVData = list(),
  metadata = data.frame(),
  sweepResult = list(),
  OTResult = list(),
  units = "micron"
)

validate_OTAnalysis(x)
}
\arguments{
\item{IVData}{A list of data frames, each restores measurements for one
transistor in three or more columns: \code{V_gate} (gate voltage),
\code{I_drain} (drain current), \code{I_gate} (gate voltage). The gate voltage sweeps
from low to high values (forward sweep), then back from high to low values
(backward sweep).}

\item{metadata}{A \code{data.frame} that restores the metadata such as substrate
material and geometry measurement for each transistor. Each row of the
\code{data.frame} represents a transistor, following the same order as in the
list of \code{IVData}. As in \code{OTDataSet}, the following information should be
provided: \code{substrate} (character), \code{length} (numeric), \code{width} (numeric),
\code{thickness} (numeric), and \code{type} (character, "p" or "n"). Based on the
dimension measurements, a new column \code{dim_ratio} (numeric) is added.}

\item{sweepResult}{A list of three lists of the same length as the number of
transistors. Each of the three lists contains data frames for sweep-level
information about the transistors. The \code{sweepData} list consists of data
frames augmented from those in \code{IVData}, with information about sweep and
calculated transconductance. The \code{QC} list provides quality measures and
corresponding outlyingness in terms of maximum drain current value,
difference between drain and gate currents, and on-off ratio and range.
The \code{estimate} list provides sweep-level summaries in terms of outlyingness,
maximum transconductance, threshold voltage, and area under the transfer
curve.}

\item{OTResult}{A list of data frames, each summarizes the analysis in terms
of outlyingness, average maximum transconductance, difference in threshold
voltage between sweeps, and difference area under the transfer curve.}

\item{units}{The unit for measurements of transistor's size ("mm", "micron",
or "nm")}

\item{x}{An \code{OTAnalysis} object}
}
\value{
An OTAnalysis object.

The validated \code{OTAnalysis} as input
}
\description{
\code{OTAnalysis} is an extended class of \code{OTDataSet} that includes extra details
about analysis results summarized at the sweep-level and transistor-level.
\code{new_OTAnalysis()} constructs an \code{OTAnalysis} object, and
\code{validate_OTAnalysis()} checks an \code{OTAnalysis} for expected format and info.

The constructor (and validator) would not typically be used by an
end user. These functions are called by the \code{OT_analyze()} function that
takes as input an \code{OTDataSet} and returns a new \code{OTAnalysis} object. They may
also be called by methods for modifying an \code{OTAnalysis} object.
}
