% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OTCore-package.R
\docType{package}
\name{OTCore-package}
\alias{OTCore}
\alias{OTCore-package}
\title{OTCore: Data Structures and Functions for Analysis of Organic Transistor Data}
\description{
Convenience functions to facilitate analysis of organic transistor measurement data. The analysis procedure handles data representation, exploration, analysis, summarization and visualization, together with helper functions for importing data and exporting analysis results.
}
\author{
\strong{Maintainer}: Tsung-Heng Tsai \email{tsai.tsungheng@gmail.com}

}
\keyword{internal}
