% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-classes.R
\docType{class}
\name{new_OTDataSet}
\alias{new_OTDataSet}
\alias{OTDataSet}
\alias{OTDataSet-class}
\alias{validate_OTDataSet}
\title{OTDataSet constructor and validator}
\usage{
new_OTDataSet(IVData = list(), metadata = data.frame(), units = "micron")

validate_OTDataSet(x)

OTDataSet(IVData = list(), metadata = data.frame(), units = "micron")
}
\arguments{
\item{IVData}{A list of data frames, each restores measurements for one
transistor in three or more columns: \code{V_gate} (gate voltage),
\code{I_drain} (drain current), \code{I_gate} (gate voltage). The gate voltage sweeps
from low to high values (forward sweep), then back from high to low values
(backward sweep).}

\item{metadata}{A \code{data.frame} that restores the metadata such as substrate
material and geometry measurement for each transistor. Each row of the
\code{data.frame} represents a transistor, following the same order as in the
list of \code{IVData}. The following information should be provided in the
columns of \code{metadata}: \code{substrate} (character), \code{length} (numeric),
\code{width} (numeric), \code{thickness} (numeric), and \code{type} (character, "p" or "n").}

\item{units}{The unit for measurements of transistor's size ("mm", "micron",
or "nm")}

\item{x}{An OTDataSet}
}
\value{
An OTDataSet object.
}
\description{
\code{OTDataSet} is a class used to restore current-voltage (I-V) measurement
data from organic transistors and their metadata. The \code{OTDataSet} class is a
list of two elements named \code{IVData} and \code{metadata}.

\code{new_OTDataSet()} constructs an \code{OTDataSet} object, and
\code{validate_OTDataSet()} checks an \code{OTDataSet} for expected format and info.
\code{OTDataSet()} is a helper function that wraps the constructor and validator.
}
\examples{
v <- c(seq(0, 0.5, by = 0.05), seq(0.45, 0, by = -0.05))
id1 <- 10^(-5) * v + rnorm(length(v), 0, 10^(-9))
id2 <- 10^(-5) * v + rnorm(length(v), 0, 10^(-9))
ig1 <- runif(length(v), 0, 10^(-9))
ig2 <- runif(length(v), 0, 10^(-9))
df1 <- data.frame(I_drain = id1, I_gate = ig1, V_gate = v)
df2 <- data.frame(I_drain = id2, I_gate = ig2, V_gate = v)
iv <- list(df1, df2)
meta <- data.frame(
  substrate = rep("A", 2),
  length = rep(10, 2),
  width = rep(5, 2),
  thickness = rep(0.1, 2),
  type = rep("p", 2)
)
ot <- OTDataSet(IVData = iv, metadata = meta)
}
