% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-classes.R
\name{OTDataSet_from_files}
\alias{OTDataSet_from_files}
\title{Create OTDataSet from CSV or Excel files}
\usage{
OTDataSet_from_files(
  filepath = NULL,
  metadata = NULL,
  required_iv = NULL,
  units = "micron"
)
}
\arguments{
\item{filepath}{A \code{data.frame} that provides the information of paths to I-V
measurement data for the transistors to be analyzed. It should contain
columns \code{path} (required), and \code{sheet} (required, if an Excel file is
provided).}

\item{metadata}{A \code{data.frame} that restores the metadata such as substrate
material and geometry measurement for each transistor. Each row of the
\code{data.frame} represents a transistor, following the same order as in the
list of \code{IVData}. The following information should be provided in the
columns of \code{metadata}: \code{substrate} (character), \code{length} (numeric),
\code{width} (numeric), \code{thickness} (numeric), and \code{type} (character, "p" or "n").}

\item{required_iv}{A named vector to specify the column names in the data
files for the required I-V variables: \code{I_drain}, \code{I_gate}, \code{V_gate}.}

\item{units}{The unit for measurements of transistor's size ("mm", "micron",
or "nm").}
}
\value{
An OTDataSet object.
}
\description{
\code{OTDataSet_from_files()} is a helper to load data from CSV or Excel files,
construct and then validate an OTDataSet object. \code{filepath} is a data frame
with columns \code{path} (required), and \code{sheet} (required, if an Excel file is
provided)
}
\examples{
# When I_drain, I_gate, and V_gate are called DrainI, GateI, and GateV,
# respectively, in the data files, we specify the correspondence as
required_iv <- c(DrainI = "I_drain", GateI = "I_gate", GateV = "V_gate")
}
