% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{summarize_sweeps}
\alias{summarize_sweeps}
\title{Summarize a transistor with sweep-level estimates}
\usage{
summarize_sweeps(x, ignore_outlier = TRUE)
}
\arguments{
\item{x}{A data frame.}

\item{ignore_outlier}{A logical value.}
}
\value{
A data frame with one row and four columns: \code{flag},
\code{V_thresh_shift}, \code{auc_shift}, \code{gm_max}.
}
\description{
Summarize a transistor with sweep-level estimates
}
